master_user renamed to base_user
